fix review generation and reviews response

+++ internal/domain/product.go
@@ -7,7 +7,7 @@ type Product struct {
 	Name        string             `json:"name" bson:"name"`
 	Description string             `json:"description" bson:"description"`
 	Price       float64            `json:"price" bson:"price"`
-	TotalRating float32            `json:"totalRating" bson:"-"`
+	TotalRating float64            `json:"totalRating" bson:"-"`
 	Categories  []Category         `json:"categories" bson:"categories"` //TODO: make category independent entity, change type to Array(ObjectID)
 }
